+++ pkg/ingester/user_state.go
@@ -16,9 +16,10 @@ import (
 
 // DiscardedSamples metric labels
 const (
-	perUserSeriesLimit     = "per_user_series_limit"
-	perMetricSeriesLimit   = "per_metric_series_limit"
-	perLabelsetSeriesLimit = "per_labelset_series_limit"
+	perUserSeriesLimit                = "per_user_series_limit"
+	perUserNativeHistogramSeriesLimit = "per_user_native_histogram_series_limit"
+	perMetricSeriesLimit              = "per_metric_series_limit"
+	perLabelsetSeriesLimit            = "per_labelset_series_limit"
 )
 
 const numMetricCounterShards = 128